+++ Kufar3/ModelsView/DeclarationModel.cs
@@ -8,11 +8,11 @@ namespace Kufar3.ModelsView
 {
     public class DeclarationModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int SubCategoryId { get; set; }
-        public int CityId { get; set; }
+        public long SubCategoryId { get; set; }
+        public long CityId { get; set; }
         public List<string> Images  { get; set; }
     }
 }
